requirements.txt
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
Werkzeug==2.0.2
